refactor : 이미지 구조 수정
