+++ src/main/java/com/nighttrip/core/global/image/ImageType.java
@@ -1,0 +1,5 @@
+package com.nighttrip.core.global.image;
+
+public enum ImageType {
+    USER, CITY, TOURLIST_SPOT, MEMO
+}
